Moved some code to utilities.js file, to clean up the index.js file
